requirements.txt
anyio==3.6.2
bcrypt==4.0.1
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.92.0
flake8==6.0.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
mangum==0.17.0
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
orjson==3.8.6
packaging==23.0
passlib==1.7.4
pathspec==0.11.0
platformdirs==3.0.0
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.5
pyflakes==3.0.1
Pygments==2.14.0
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
rich==13.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
